backend/requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
python-dotenv==1.2.1
requests==2.32.5
urllib3==2.6.2
fastapi
uvicorn
sqlmodel
aiosqlite
httpx
openai
google-genai
python-dotenv
pydantic
pyyaml
bandit
flake8
flake8-json

# Phase 1: Vector DB & Code Analysis
qdrant-client
tree-sitter==0.21.3
tree-sitter-languages
networkx
redis
tiktoken
asyncpg
psycopg2-binary
PyJWT
groq
